Create ball and bumper class, start testing
In the interests of allowing other team members to develop using new ball and
bumper classes, this commit separates out ball and bumper files and only initiates
the beginning of testing.
