+++ test/ball_test.js
@@ -1,0 +1,4 @@
+const chai = require('chai')
+const assert = chai.assert;
+
+const Ball = require("../lib/ball")
