+++ test/index.js
@@ -1,0 +1,1 @@
+require('./ball_test.js')
